[front]add breadcrumb notes, external references & tsx(#8274)
